Move patches to engine specific
